+++ lib/Peast/Selector/Node/Part/Pseudo.php
@@ -13,6 +13,8 @@ namespace Peast\Selector\Node\Part;
  * Selector pseudo part base class
  * 
  * @author Marco Marchiò <marco.mm89@gmail.com>
+ *
+ * @abstract
  */
 abstract class Pseudo extends Part
 {